# perf: Perform `db.set_value` with single query only