Concludes https://github.com/frappe/frappe/pull/15560  by making all set_value calls to non-single doctypes run with 1 query only. 


#### Explanation:

Previously we were fetching all the primary keys which need to be updated and then updating separately. 

Before:
```
names = frappe.db.get_values(... filters)
frappe.qb.update(fields).where(table.name.isin(names))
```

After:
```
frappe.qb.update(fields).where(filters)
```


QB can already do single update query without much fuss. The last reason for keeping this appears to be `for update` which locks the rows before updating them. 

#### Why remove `for_update` completely?

This was added to make sure that N+1 queries that happen don't cause inconsistencies during the time queries are ran. Now that set_value is 1-1 translated to single `UPDATE` query there's no point in locking rows before updating them. `UPDATE` locks rows automatically. 


Before:
`1.02 ms ± 166 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)`

After:
`530 µs ± 144 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)`


Roughly 2x faster. This will be much better on queries that end up affecting more than 1 row. 

TODO:

- [x] better cache eviction when dn is not a filter
- [ ] ~~evict only doctype~~ (deferred, needs design change in cache keys)

ERPNext CI: https://github.com/frappe/erpnext/actions/runs/3197465411 ✅